# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /time-logs:
    get:
      operationId: listTimeLogs
      tags: [payroll]
      summary: List time logs, optionally filtered by employee and/or date range
      description: >-
        Powers the Time Tracking hub (PRD Section 4) — Date/Employee/Role/Hours table.
        `start`/`end` (both `YYYY-MM-DD`, both optional, inclusive) filter on `date`.
      parameters:
        - name: employeeId
          in: query
          required: false
          schema:
            type: integer
        - name: start
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Matching time logs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeLogResult"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTimeLog
      tags: [payroll]
      summary: Manually log hours for an employee/role/day
      description: >-
        Manual entry only for v1 (PRD Open Question #11's decided default: "manual
        entry with admin approval," no clock-in/out). Always created unapproved — see
        `POST /time-logs/{id}/approve`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimeLogRequest"
      responses:
        "201":
          description: Time log created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeLogResult"
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /time-logs/{id}/approve:
    post:
      operationId: approveTimeLog
      tags: [payroll]
      summary: Approve a time log
      description: >-
        Admin review/lock step ahead of a payroll run (PRD Section 4) — only approved
        logs feed a payroll run's `hourly_pay`.
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Approved time log.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeLogResult"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No time log with this id exists in the current organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /time-logs/{id}:
    delete:
      operationId: deleteTimeLog
      tags: [payroll]
      summary: Delete a time log
      description: Removes a mis-entered manual log — typically before it's approved.
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Time log deleted.
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No time log with this id exists in the current organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /time-off-requests:
    get:
      operationId: listTimeOffRequests
      tags: [payroll]
      summary: List time-off requests, optionally filtered by status and/or employee
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, approved, denied]
        - name: employeeId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Matching time-off requests.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeOffRequestResult"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTimeOffRequest
      tags: [payroll]
      summary: Submit a time-off request
      description: >-
        Created with `status: pending` (PRD Section 5). Approved time off is
        informational only in v1 (Open Question #12) — it never blocks or warns
        against new booking assignments in that date range.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimeOffRequestRequest"
      responses:
        "201":
          description: Time-off request created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeOffRequestResult"
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /time-off-requests/{id}/review:
    post:
      operationId: reviewTimeOffRequest
      tags: [payroll]
      summary: Approve or deny a time-off request
      description: >-
        Matches the "Approve Time Off Request" confirmation modal (PRD Section 5).
        `reviewedByEmployeeId` is optional — v1 has no reliable mapping from the
        authenticated admin user to an `employees` row (no employee login yet).
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewTimeOffRequestRequest"
      responses:
        "200":
          description: Reviewed time-off request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeOffRequestResult"
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No time-off request with this id exists in the current organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings, optionally within a date range and/or for one client
      description: >-
        Powers the calendar view (FR-10) — `start`/`end` (both `YYYY-MM-DD`, both
        optional, inclusive) filter on `date` rather than fetching every booking ever
        made. Omitting both returns every booking for the organization. `clientId`
        (Section 6.2 of PRD_DetailHub_Signup_Copy_and_Packages_Hardening.md) further
        narrows the result to bookings for that client — it composes with (never
        replaces) the organization scope every request is already limited to, so a
        `clientId` belonging to a different organization returns an empty array
        rather than another organization's booking.
      parameters:
        - name: start
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Matching bookings.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResult"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      description: >-
        `packageIds` and `employeeSplit` fully describe the booking's line items and
        employee revenue split in one call (a `booking_packages`/`employee_splits`
        join row is written per entry). `employeeSplit` may be empty (a booking can be
        created before an employee is assigned), but if non-empty every percentage
        must be in `(0, 100]` and the set must sum to exactly 100.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResult"
        "400":
          description: >-
            Invalid request body, an invalid `employeeSplit` (see this operation's
            description), or a `clientId`/`packageIds`/`employeeSplit[].employeeId`
            that doesn't reference a real row in this organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    EmployeeSplit:
      type: object
      description: >-
        Matches `artifacts/detail-hub/src/lib/mock-data.ts`'s `EmployeeSplit`
        interface. `Booking.employeeIds` is folded into this (see `BookingResult`) —
        the split's `employeeId` list already implies assignment, so a separate raw
        id list would just be redundant data that could drift.
      properties:
        employeeId:
          type: integer
        percentage:
          type: number
          minimum: 0
          maximum: 100
      required:
        - employeeId
        - percentage
...
    TimeLogResult:
      type: object
      description: >-
        `docs/prds/PRD_DetailHub_Payroll_Module.md` Section 2.4. `source`/
        `linkedBookingId` stay schema-supported for a future auto-derive-from-booking
        path, but every row created through this API today is `manual_entry` with
        `linkedBookingId: null` (Open Question #11's decided v1 default).
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        roleName:
          type: string
        date:
          type: string
          format: date
        hours:
          type: number
        source:
          type: string
          enum: [manual_entry, derived_from_booking]
        linkedBookingId:
          type: integer
          nullable: true
        approved:
          type: boolean
      required:
        - id
        - employeeId
        - roleName
        - date
        - hours
        - source
        - linkedBookingId
        - approved
    CreateTimeLogRequest:
      type: object
      description: >-
        Manual entry only — always created with `source: manual_entry`,
        `approved: false`.
      properties:
        employeeId:
          type: integer
        roleName:
          type: string
          minLength: 1
        date:
          type: string
          format: date
        hours:
          type: number
          minimum: 0
      required:
        - employeeId
        - roleName
        - date
        - hours
    TimeOffRequestResult:
      type: object
      description: >-
        `docs/prds/PRD_DetailHub_Payroll_Module.md` Section 2.5. Per Open Question
        #12's decided default, an `approved` request is informational only — this API
        never blocks/warns against new booking assignments in that date range.
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, approved, denied]
        requestedAt:
          type: string
          format: date-time
        reviewedBy:
          type: integer
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        note:
          type: string
          nullable: true
      required:
        - id
        - employeeId
        - startDate
        - endDate
        - status
        - requestedAt
        - reviewedBy
        - reviewedAt
    CreateTimeOffRequestRequest:
      type: object
      properties:
        employeeId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        note:
          type: string
          nullable: true
      required:
        - employeeId
        - startDate
        - endDate
    ReviewTimeOffRequestRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, denied]
        reviewedByEmployeeId:
          type: integer
          nullable: true
          description: >-
            Optional — no reliable mapping from the authenticated admin user to an
            `employees` row exists in v1 (no employee login yet).
      required:
        - status
...
    BookingResult:
      type: object
      description: >-
        Field names/shape match `artifacts/detail-hub/src/lib/mock-data.ts`'s
        `Booking` interface (`date`, `startTime`, `address`, not the master PRD's
        `scheduledDate`/`scheduledStartTime`/`serviceAddress` — the mock interface is
        the source of truth per Section 7), plus `createdAt`/`updatedAt`/`createdBy`.
        `employeeIds` is derived from `employeeSplit` for convenience, not stored
        separately. (`gasMeterStatus`/`weatherSnapshot` were removed per FR-9 of
        PRD_DetailHub_Signup_Copy_and_Packages_Hardening.md — dead placeholder
        columns never populated by a real integration; FR-11 says they come back for
        real later, see Fuel_Gauge_PRD.md.) `paymentMethod`/`paymentReference`/
        `paymentRecordedAt`/`refundStatus`/`refundReference` match
        `PRD_DetailHub_Payment_Methods.md` Section 7 exactly (superseding the earlier,
        rougher `paymentMethod`/`paymentNote` shape this API had before that PRD was
        reviewed — see `../../lib/db/src/schema/bookings.ts` for why) — settable only
        via `POST /bookings/{id}/payment` and `POST /bookings/{id}/refund`, never
        through `POST`/`PATCH /bookings`.
      properties:
        id:
          type: integer
        clientId:
          type: integer
        packageIds:
          type: array
          items:
            type: integer
        employeeIds:
          type: array
          items:
            type: integer
          description: Derived from `employeeSplit` (every `employeeId` present in the split).
        employeeSplit:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeSplit"
        date:
          type: string
          format: date
        startTime:
          type: string
          description: "24-hour HH:MM, e.g. \"09:00\"."
        address:
          type: string
        depositAmount:
          type: number
        parkingCost:
          type: number
        status:
          type: string
          enum: [confirmed, pending, completed, cancelled, no-show]
        notes:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
          enum: [zelle, venmo, cash, credit_card]
        paymentReference:
          type: string
          nullable: true
        paymentRecordedAt:
          type: string
          format: date-time
          nullable: true
        refundStatus:
          type: string
          nullable: true
          enum: [none, requested, completed]
        refundReference:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          description: id of the admin/owner user who created this booking.
      required:
        - id
        - clientId
        - packageIds
        - employeeIds
        - employeeSplit
        - date
        - startTime
        - address
        - depositAmount
        - parkingCost
        - status
        - createdAt
        - updatedAt
        - createdBy
    CreateBookingRequest:
      type: object
      properties:
        clientId:
          type: integer
        packageIds:
          type: array
          items:
            type: integer
          default: []
        employeeSplit:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeSplit"
          default: []
        date:
          type: string
          format: date
        startTime:
          type: string
        address:
          type: string
          minLength: 1
        depositAmount:
          type: number
          minimum: 0
        parkingCost:
          type: number
          minimum: 0
        status:
          type: string
          enum: [confirmed, pending, completed, cancelled, no-show]
        notes:
          type: string
          nullable: true
      required:
        - clientId
        - date
        - startTime
        - address
        - depositAmount
        - parkingCost
        - status
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable detail, when available.
      required:
        - error