# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingRequestStatus } from './createBookingRequestStatus';
import type { EmployeeSplit } from './employeeSplit';

export interface CreateBookingRequest {
  clientId: number;
  packageIds?: number[];
  employeeSplit?: EmployeeSplit[];
  date: Date;
  startTime: string;
  /** @minLength 1 */
  address: string;
  /** @minimum 0 */
  depositAmount: number;
  /** @minimum 0 */
  parkingCost: number;
  status: CreateBookingRequestStatus;
  notes?: string | null;
}