# lib/api-zod/src/generated/types/createTimeOffRequestRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTimeOffRequestRequest {
  employeeId: number;
  startDate: Date;
  endDate: Date;
  note?: string | null;
}