# lib/api-zod/src/generated/types/reviewTimeOffRequestRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewTimeOffRequestRequestStatus } from './reviewTimeOffRequestRequestStatus';

export interface ReviewTimeOffRequestRequest {
  status: ReviewTimeOffRequestRequestStatus;
  /** Optional — no reliable mapping from the authenticated admin user to an `employees` row exists in v1 (no employee login yet). */
  reviewedByEmployeeId?: number | null;
}